requirements.txt
django==4.1
django-widget-tweaks==1.4.12
loguru==0.6.0
Pillow==9.2.0
gunicorn==20.1.0
whitenoise==6.1.0
six==1.16.0 
python-dotenv==0.20.0
psycopg2
requests~=2.28.2
djangorestframework